Lab Work/17. Pydantic MCP server/pyproject.toml
[project]
name = "9-3-pydantic-ai-agent-mcp-tool"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai-slim[groq,mcp]>=1.0.1",
    "python-dotenv>=1.1.1",
]
